+++ internal/server/server.go
@@ -8,6 +8,7 @@ import (
 	"github.com/hashicorp/vault-csi-provider/internal/config"
 	"github.com/hashicorp/vault-csi-provider/internal/provider"
 	"github.com/hashicorp/vault-csi-provider/internal/version"
+	"k8s.io/client-go/kubernetes"
 	pb "sigs.k8s.io/secrets-store-csi-driver/provider/v1alpha1"
 )
 
@@ -19,12 +20,14 @@ var (
 type Server struct {
 	logger      hclog.Logger
 	flagsConfig config.FlagsConfig
+	k8sClient   kubernetes.Interface
 }
 
-func NewServer(logger hclog.Logger, flagsConfig config.FlagsConfig) *Server {
+func NewServer(logger hclog.Logger, flagsConfig config.FlagsConfig, k8sClient kubernetes.Interface) *Server {
 	return &Server{
 		logger:      logger,
 		flagsConfig: flagsConfig,
+		k8sClient:   k8sClient,
 	}
 }
 
@@ -42,7 +45,7 @@ func (s *Server) Mount(ctx context.Context, req *pb.MountRequest) (*pb.MountResp
 		return nil, err
 	}
 
-	provider := provider.NewProvider(s.logger.Named("provider"))
+	provider := provider.NewProvider(s.logger.Named("provider"), s.k8sClient)
 	resp, err := provider.HandleMountRequest(ctx, cfg, s.flagsConfig)
 	if err != nil {
 		return nil, fmt.Errorf("error making mount request: %w", err)