+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ## Unreleased
 
+CHANGES:
+
+* Duplicate object names now trigger an error instead of silently overwriting files. [[GH-148](https://github.com/hashicorp/vault-csi-provider/pull/148)]
+
+IMPROVEMENTS:
+
+* Secret versions are now reported as a hash of their input and contents instead of hardcoded to 0. [[GH-148](https://github.com/hashicorp/vault-csi-provider/pull/148)]
+
 ## 1.1.0 (April 26th, 2022)
 
 IMPROVEMENTS: